Fix records not being removed on deletion
Resolve an issue where modifications to the Class field were taking
place on the original record. This affected the ability to later
delete that record due to entries.contains() returning false because
the record being compared no longer matched the stored record due to
the previous modification of the class field. This commit modifies the
code to make deep copies of response records so that modifications to
the Class field do not take place on the original record.

This commit also fixes the removal of hostnames with one record entry.