Fix missing Docker image labels
The labels that were intended to be on the Docker image have been
missing since the initial 0.1.0 release. This is because the labels
were being incorrectly applied in the build stage instead of to the
final image.

This commit modifies the Dockerfile so that the labels are correctly
configured on the final image.